create NavGraphBuilderExt
