+++ app/src/main/kotlin/com/example/ui/MainActivity.kt
@@ -18,10 +18,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
+import com.example.addAddTodo
 import com.example.addCatalog
-import com.example.ui.home.detail.addDetail
-import com.example.ui.home.favorite.addFavorite
-import com.example.ui.home.grid.addGrid
+import com.example.addDetail
+import com.example.addFavorite
+import com.example.addGrid
+import com.example.addLogIn
+import com.example.addSignUp
+import com.example.addTodo
+import com.example.addUpdateTodo
 import com.example.ui.todo.TodoViewModel
 import com.example.ui.todo.todo.*
 import com.example.ui.utils.Routes
