+++ app/src/main/kotlin/com/example/ui/home/favorite/FavoriteScreen.kt
@@ -1,17 +1,25 @@
 package com.example.ui.home.favorite
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.Divider
 import androidx.compose.material.ScaffoldState
 import androidx.compose.material.icons.Icons
-import androidx.compose.material3.Text
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.FavoriteBorder
+import androidx.compose.material3.Divider
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -23,9 +31,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
-import androidx.navigation.compose.composable
 import com.example.composetrainapp.R
 import com.example.model.Character
 import com.example.ui.home.RickMortyViewModel
@@ -35,18 +41,6 @@ import com.example.ui.utils.compose.FullScreenLoadingIndicator
 import com.example.ui.utils.compose.ToggleButton
 import com.example.ui.utils.compose.TrainAppImage
 import com.example.ui.utils.showSnackBar
-
-fun NavGraphBuilder.addFavorite(
-    scaffoldState: ScaffoldState,
-    navController: NavHostController,
-) {
-    composable(route = Routes.Favorite.route) {
-        FavoriteScreen(
-            scaffoldState = scaffoldState,
-            navController = navController,
-        )
-    }
-}
 
 @Composable
 fun FavoriteScreen(
