+++ app/src/main/kotlin/com/example/ui/todo/todo/AddTodoScreen.kt
@@ -1,28 +1,34 @@
 package com.example.ui.todo.todo
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Button
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.CheckboxDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
 import com.example.ui.todo.TodoViewModel
 import com.example.ui.todo.checkIsTodoBodyValid
 import com.example.ui.todo.checkIsTodoTitleValid
 import com.example.ui.utils.Routes
-
-fun NavGraphBuilder.addAddTodo(navController: NavController) {
-    composable(route = Routes.CreateTodo.route) {
-        AddTodoScreen(modifier = Modifier.padding(4.dp), navController)
-    }
-}
 
 @Composable
 fun AddTodoScreen(
