+++ app/src/main/kotlin/com/example/ui/todo/todo/TodoLogInScreen.kt
@@ -29,9 +29,7 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
-import androidx.navigation.compose.composable
 import com.example.composetrainapp.R
 import com.example.data.utils.Result
 import com.example.model.User
@@ -43,12 +41,6 @@ import com.example.ui.todo.TodoViewModel
 import com.example.ui.utils.Routes
 import com.example.ui.utils.compose.FullScreenLoadingIndicator
 import showToast
-
-fun NavGraphBuilder.addLogIn(navController: NavHostController) {
-    composable(route = Routes.LogIn.route) {
-        TodoLogInScreen(navController)
-    }
-}
 
 @Composable
 fun TodoLogInScreen(
