+++ feature/catalog/src/main/kotlin/com/example/feature/catalog/CatalogScreen.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListScope
-import androidx.compose.material3.Divider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -15,8 +14,6 @@ import showToast
 fun CatalogScreen() {
     val context = LocalContext.current
     CatalogContent {
-        item { Divider(color = MaterialTheme.colorScheme.onBackground) }
-
         catalogItem(label = "label 1") {
             context.showToast("label 1")
         }
